% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blur_pkgs_images}
\alias{blur_pkgs_images}
\title{Blur image from selected list of packages}
\usage{
blur_pkgs_images(pkgs, path = "inst/www/blurred")
}
\arguments{
\item{pkgs}{Vector of package names.}

\item{path}{Path to images.}
}
\value{
Side effect: create blurred png files of input images.
}
\description{
Blur image from selected list of packages
}
\keyword{internal}
