% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_and_update_pkgs_hex}
\alias{check_and_update_pkgs_hex}
\title{Update package description and images}
\usage{
check_and_update_pkgs_hex(path = "inst/www/blurred")
}
\arguments{
\item{path}{Path where current images are located}
}
\description{
Query GitHub for package names. Compare
to local package folder and update only elements
that are missing in the local folder. Images are blurred and description
file is modified to add the new packages if any.
}
\keyword{internal}
