% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_rand_pkg_authors}
\alias{extract_rand_pkg_authors}
\title{Generate random package author list}
\usage{
extract_rand_pkg_authors(pkg)
}
\arguments{
\item{pkg}{Package to search for in r-unviverse API.}
}
\value{
list containing package name, authors and maintainer.
}
\description{
One of the author is the maintainer
}
\keyword{internal}
