% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_descriptions}
\alias{get_descriptions}
\title{Get description files}
\usage{
get_descriptions(pkgs = get_pkgs_names())
}
\arguments{
\item{pkgs}{Vector of package names to write the description.
Default to \link{get_pkgs_names}.}
}
\description{
That's actually just a subset of the description.
This function is used for pre-processing data and not intended
to be used by the app directly. The app uses the descriptions file
located in the path parameter.
}
\keyword{internal}
