% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkg_from_runiverse}
\alias{get_pkg_from_runiverse}
\title{Get package info on r-universe}
\usage{
get_pkg_from_runiverse(pkg)
}
\arguments{
\item{pkg}{Package name as string.}
}
\value{
A list of metadata.
}
\description{
Get package info on r-universe
}
\keyword{internal}
