% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkgs_names}
\alias{get_pkgs_names}
\title{Get package names from GitHub}
\usage{
get_pkgs_names(
  endpoint = "/repos/rstudio/hex-stickers/git/trees/main?recursive=1"
)
}
\arguments{
\item{url}{Location to read from.}
}
\value{
Vector package names.
}
\description{
Get package names from GitHub
}
\keyword{internal}
