% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_authors}
\alias{pad_authors}
\title{Pad a list with random authors}
\usage{
pad_authors(descs)
}
\description{
Some packages do not have four authors in the DESCRIPTION file that can be
used for the quiz. This function checks the number of authors for a given
DESCRIPTION file and pads the the number of authors so that it always
equals four.
}
