% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_pkg_authors}
\alias{process_pkg_authors}
\title{Only returns given and family name of a package author}
\usage{
process_pkg_authors(authors)
}
\arguments{
\item{authors}{Vector of authors returned by the desc package.}
}
\value{
character vector
}
\description{
Only returns given and family name of a package author
}
\keyword{internal}
