% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_descriptions}
\alias{write_descriptions}
\title{Write description file}
\usage{
write_descriptions(
  descs = get_descriptions(),
  path = "inst/www/descriptions.rds"
)
}
\arguments{
\item{descs}{Vector of package names to write the description. Default to \link{get_descriptions}.}

\item{path}{Path to save descriptions list to.}
}
\description{
Save the description file to disk. Where the number of possible of authors is
less than four, the author list should be padded with random authors from
other packages.
}
\keyword{internal}
